(not working) Updating parser to support foreach

+++ Midas/src/c-compiler/function/expressions/shared/shared.cpp
@@ -266,7 +266,7 @@ LLVMValueRef makeConstExpr(FunctionState* functionState, LLVMBuilderRef builder,
   return LLVMBuildLoad(builder, localAddr, "");
 }
 
-LLVMValueRef makeConstIntExpr(FunctionState* functionState, LLVMBuilderRef builder, LLVMTypeRef type, int value) {
+LLVMValueRef makeConstIntExpr(FunctionState* functionState, LLVMBuilderRef builder, LLVMTypeRef type, int64_t value) {
   return makeConstExpr(functionState, builder, LLVMConstInt(type, value, false));
 }
