+++ Midas/src/c-compiler/function/expressions/shared/shared.h
@@ -155,7 +155,7 @@ Ref buildInterfaceCall(
     int virtualParamIndex);
 
 
-LLVMValueRef makeConstIntExpr(FunctionState* functionState, LLVMBuilderRef builder, LLVMTypeRef type, int value);
+LLVMValueRef makeConstIntExpr(FunctionState* functionState, LLVMBuilderRef builder, LLVMTypeRef type, int64_t value);
 
 LLVMValueRef makeConstExpr(
     FunctionState* functionState, LLVMBuilderRef builder, LLVMValueRef constExpr);
